moved leaderboard page to a component
